ggpubr-package {ggpubr}	R Documentation

ggpubr: 'ggplot2' Based Publication Ready Plots

Description

The 'ggplot2' package is excellent and flexible for elegant data visualization in R. However the default generated plots requires some formatting before we can send them for publication. Furthermore, to customize a 'ggplot', the syntax is opaque and this raises the level of difficulty for researchers with no advanced R programming skills. 'ggpubr' provides some easy-to-use functions for creating and customizing 'ggplot2'- based publication ready plots.

Details

General resources:

ggpubr documentation

ggpubr tutorials

Package options

ggpubr.parse_aes

logical indicating whether to parse or not aesthetics variables names. Default is TRUE. For example, if you want ggpubr to handle non-standard column names, like A-A, without parsing, then set this option to FALSE using options(ggpubr.parse_aes = FALSE).

ggpubr.null_device

A function that creates an appropriate null device. These include: cowplot::pdf_null_device, cowplot::png_null_device, cowplot::cairo_null_device and cowplot::agg_null_device. Default is cowplot::pdf_null_device. This is used in function like as_ggplot(), which needs to open a graphics device to render ggplot objects into grid graphics objects. This function is used to open null device for avoiding the display of unnecessary blank page when calling ggarrange() or as_ggplot()

Author(s)

Maintainer: Alboukadel Kassambara alboukadel.kassambara@gmail.com

Other contributors:

Laszlo Erdey erdey.laszlo@econ.unideb.hu (Faculty of Economics and Business, University of Debrecen, Hungary) [contributor]

See Also

Useful links:

https://rpkgs.datanovia.com/ggpubr/

Report bugs at https://github.com/kassambara/ggpubr/issues

[Package ggpubr version 0.6.3 Index]