Comparison of Means

Performs one or multiple mean comparisons.

compare_means(formula, data, method = "wilcox.test", paired = FALSE,
  group.by = NULL, ref.group = NULL, symnum.args = list(),
  p.adjust.method = "holm", ...)

Arguments

formula
a formula of the form x ~ group where x is a numeric variable giving the data values and group is a factor with one or multiple levels giving the corresponding groups. For example, formula = TP53 ~ cancer_group.

It's also possible to perform the test for multiple response variables at the same time. For example, formula = c(TP53, PTEN) ~ cancer_group.

data
a data.frame containing the variables in the formula.

method
the type of test. Default is wilcox.test. Allowed values include:

t.test (parametric) and wilcox.test (non-parametric). Perform comparison between two groups of samples. If the grouping variable contains more than two levels, then a pairwise comparison is performed.

anova (parametric) and kruskal.test (non-parametric). Perform one-way ANOVA test comparing multiple groups.

paired
a logical indicating whether you want a paired test. Used only in t.test and in wilcox.test.

group.by
a character vector containing the name of grouping variables.

ref.group
a character string specifying the reference group. If specified, for a given grouping variable, each of the group levels will be compared to the reference group (i.e. control group).

ref.group can be also ".all.". In this case, each of the grouping variable levels is compared to all (i.e. basemean).

symnum.args
a list of arguments to pass to the function symnum for symbolic number coding of p-values. For example, symnum.args <- list(cutpoints = c(0, 0.0001, 0.001, 0.01, 0.05, 1), symbols = c("****", "***", "**", "*", "ns")).

In other words, we use the following convention for symbols indicating statistical significance:

ns: p > 0.05

*: p <= 0.05

**: p <= 0.01

***: p <= 0.001

****: p <= 0.0001

p.adjust.method
method for adjusting p values (see p.adjust). Has impact only in a situation, where multiple pairwise tests are performed; or when there are multiple grouping variables. Allowed values include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". If you don't want to adjust the p value (not recommended), use p.adjust.method = "none".

Note that, when the formula contains multiple variables, the p-value adjustment is done independently for each variable.

...
Other arguments to be passed to the test function.

Value

return a data frame with the following columns:

.y.: the y variable used in the test.

group1,group2: the compared groups in the pairwise tests. Available only when method = "t.test" or method = "wilcox.test".

p: the p-value.

p.adj: the adjusted p-value. Default for p.adjust.method = "holm".

p.format: the formatted p-value.

p.signif: the significance level.

method: the statistical test used to compare groups.

Examples

# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth

# One-sample test
#:::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ 1, df, mu = 0)

#> [90m# A tibble: 1 x 8[39m
#>   .y.   group1 group2            p    p.adj p.format p.signif method  
#>   [3m[90m<chr>[39m[23m  [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m         [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m   
#> [90m1[39m len        1 null model 1.66[90me[39m[31m-11[39m 1.70[90me[39m[31m-11[39m 1.7e-11  ****     Wilcoxon

# Two-samples unpaired test
#:::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ supp, df)

#> [90m# A tibble: 1 x 8[39m
#>   .y.   group1 group2      p p.adj p.format p.signif method  
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m   [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m   
#> [90m1[39m len   OJ     VC     0.064[4m5[24m 0.064 0.064    ns       Wilcoxon

# Two-samples paired test
#:::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ supp, df, paired = TRUE)

#> [90m# A tibble: 1 x 8[39m
#>   .y.   group1 group2       p  p.adj p.format p.signif method  
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m   
#> [90m1[39m len   OJ     VC     0.004[4m3[24m[4m1[24m 0.004[4m3[24m 0.0043   **       Wilcoxon

# Compare supp levels after grouping the data by "dose"
#::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ supp, df, group.by = "dose")

#> [90m# A tibble: 3 x 9[39m
#>    dose .y.   group1 group2       p p.adj p.format p.signif method  
#>   [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m   
#> [90m1[39m   0.5 len   OJ     VC     0.023[4m2[24m  0.046 0.023    *        Wilcoxon
#> [90m2[39m   1   len   OJ     VC     0.004[4m0[24m[4m3[24m 0.012 0.004    **       Wilcoxon
#> [90m3[39m   2   len   OJ     VC     1       1     1.000    ns       Wilcoxon

# pairwise comparisons
#::::::::::::::::::::::::::::::::::::::::
# As dose contains more thant two levels ==>
# pairwise test is automatically performed.
compare_means(len ~ dose, df)

#> [90m# A tibble: 3 x 8[39m
#>   .y.   group1 group2            p      p.adj p.format p.signif method  
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m         [3m[90m<dbl>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m   
#> [90m1[39m len   0.5    1      0.000[4m0[24m[4m0[24m[4m7[24m02   0.000[4m0[24m[4m1[24m[4m4[24m   7.0e-06  ****     Wilcoxon
#> [90m2[39m len   0.5    2      0.000[4m0[24m[4m0[24m[4m0[24m084[4m1[24m 0.000[4m0[24m[4m0[24m[4m0[24m25 8.4e-08  ****     Wilcoxon
#> [90m3[39m len   1      2      0.000[4m1[24m[4m7[24m[4m7[24m     0.000[4m1[24m[4m8[24m    0.00018  ***      Wilcoxon

# Comparison against reference group
#::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ dose, df, ref.group = "0.5")

#> [90m# A tibble: 2 x 8[39m
#>   .y.   group1 group2            p      p.adj p.format p.signif method  
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m         [3m[90m<dbl>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m   
#> [90m1[39m len   0.5    1      0.000[4m0[24m[4m0[24m[4m7[24m02   0.000[4m0[24m[4m0[24m[4m7[24m   7.0e-06  ****     Wilcoxon
#> [90m2[39m len   0.5    2      0.000[4m0[24m[4m0[24m[4m0[24m084[4m1[24m 0.000[4m0[24m[4m0[24m[4m0[24m17 8.4e-08  ****     Wilcoxon

# Comparison against all
#::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ dose, df, ref.group = ".all.")

#> [90m# A tibble: 3 x 8[39m
#>   .y.   group1 group2         p    p.adj p.format p.signif method  
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m   
#> [90m1[39m len   .all.  0.5    0.000[4m0[24m[4m5[24m[4m0[24m8 0.000[4m1[24m[4m5[24m[4m0[24m 5.1e-05  ****     Wilcoxon
#> [90m2[39m len   .all.  1      0.764     0.76     0.76404  ns       Wilcoxon
#> [90m3[39m len   .all.  2      0.000[4m1[24m[4m7[24m[4m9[24m  0.000[4m3[24m[4m6[24m  0.00018  ***      Wilcoxon

# Anova and kruskal.test
#::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ dose, df, method = "anova")

#> [90m# A tibble: 1 x 6[39m
#>   .y.          p    p.adj p.format p.signif method
#>   [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m 
#> [90m1[39m len   9.53[90me[39m[31m-16[39m 9.50[90me[39m[31m-16[39m 9.5e-16  ****     Anova 

compare_means(len ~ dose, df, method = "kruskal.test")

#> [90m# A tibble: 1 x 6[39m
#>   .y.               p        p.adj p.format p.signif method        
#>   [3m[90m<chr>[39m[23m         [3m[90m<dbl>[39m[23m        [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m         
#> [90m1[39m len   0.000[4m0[24m[4m0[24m[4m0[24m001[4m4[24m[4m8[24m 0.000[4m0[24m[4m0[24m[4m0[24m001[4m5[24m 1.5e-09  ****     Kruskal-Wallis