Comparison of Means

Performs one or multiple mean comparisons.

compare_means(formula, data, method = "wilcox.test", group.by = NULL,
  ref.group = NULL, symnum.args = list(), ...)

Arguments

formula

a formula of the form x ~ group where x is a numeric variable giving the data values and group is a factor with one or multiple levels giving the corresponding groups. For example, formula = TP53 ~ cancer_group. It's also possible to perform the test for multiple response variables at the same time. For example, formula = c(TP53, PTEN) ~ cancer_group.

data

a data.frame containing the variables in the formula.

method

the type of test. Default is wilcox.test. Allowed values include:

t.test (parametric) and wilcox.test (non-parametric). Perform comparison between two groups of samples. If the grouping variable contains more than two levels, then a pairwise comparison is performed.

anova (parametric) and kruskal.test (non-parametric). Perform one-way ANOVA test comparing multiple groups.

group.by

a character vector containing the name of grouping variables.

ref.group

a character string specifying the reference group. If specified, for a given grouping variable, each of the group levels will be compared to the reference group (i.e. control group). ref.group can be also ".all.". In this case, each of the grouping variable levels is compared to all (i.e. basemean).

symnum.args

a list of arguments to pass to the function symnum for symbolic number coding of p-values. For example, symnum.args <- list(list(cutpoints = c(0, 0.0001, 0.001, 0.01, 0.05, 0.1, 1), symbols = c("****", "***", "**", "*", "+", "NS"))).

...

Other arguments to be passed to the test function.

Value

return a data frame.

Examples

# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth

# One-sample test
#:::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ 1, df, mu = 0)

#>   .y. group1     group2      p.value p.value.format p.value.signif      method
#> 1 len      1 null model 1.664007e-11        1.7e-11           **** wilcox.test

# Two-samples unpaired test
#:::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ supp, df)

#>   .y. group1 group2    p.value p.value.format p.value.signif      method
#> 1 len     OJ     VC 0.06449067          0.064              + wilcox.test

# Two-samples paired test
#:::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ supp, df, paired = TRUE)

#>   .y. group1 group2     p.value p.value.format p.value.signif      method
#> 1 len     OJ     VC 0.004312554         0.0043             ** wilcox.test

# Compare supp levels after grouping the data by "dose"
#::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ supp, df, group.by = "dose")

#> # A tibble: 3 x 8
#>    dose   .y. group1 group2     p.value p.value.format p.value.signif
#>   <dbl> <chr>  <chr>  <chr>       <dbl>          <chr>          <chr>
#> 1   0.5   len     OJ     VC 0.023186427          0.023              *
#> 2   1.0   len     OJ     VC 0.004030367          0.004             **
#> 3   2.0   len     OJ     VC 1.000000000          1.000             NS
#> # ... with 1 more variables: method <chr>

# pairwise comparisons
#::::::::::::::::::::::::::::::::::::::::
# As dose contains more thant two levels ==>
# pairwise test is automatically performed.
compare_means(len ~ dose, df)

#>   .y. group1 group2      p.value p.value.format p.value.signif      method
#> 1 len    0.5      1 1.404171e-05        1.4e-05           **** wilcox.test
#> 2 len    0.5      2 2.521934e-07        2.5e-07           **** wilcox.test
#> 3 len      1      2 1.772382e-04        0.00018            *** wilcox.test

# Comparison against reference group
#::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ dose, df, ref.group = "0.5")

#>   .y. group1 group2      p.value p.value.format p.value.signif      method
#> 1 len    0.5      1 1.404171e-05        1.4e-05           **** wilcox.test
#> 2 len    0.5      2 2.521934e-07        2.5e-07           **** wilcox.test

# Comparison against all
#::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ dose, df, ref.group = ".all.")

#>   .y. group1 group2      p.value p.value.format p.value.signif      method
#> 1 len  .all.    0.5 0.0002031515        0.00020            *** wilcox.test
#> 2 len  .all.      1 0.7640429386        0.76404             NS wilcox.test
#> 3 len  .all.      2 0.0005317147        0.00053            *** wilcox.test

# Anova and kruskal.test
#::::::::::::::::::::::::::::::::::::::::
compare_means(len ~ dose, df, method = "anova")

#>   .y.      p.value p.value.format p.value.signif method
#> 1 len 9.532727e-16        9.5e-16           ****  anova

compare_means(len ~ dose, df, method = "kruskal.test")

#>   .y.      p.value p.value.format p.value.signif       method
#> 1 len 1.475207e-09        1.5e-09           **** kruskal.test