Descriptive statistics by groups

Computes descriptive statistics by groups for a measure variable.

desc_statby(data, measure.var, grps, ci = 0.95)

Arguments

data

a data frame.

measure.var

the name of a column containing the variable to be summarized.

grps

a character vector containing grouping variables; e.g.: grps = c("grp1", "grp2")

ci

the percent range of the confidence interval (default is 0.95).

Value

A data frame containing descriptive statistics, such as:

length: the number of elements in each group

min: minimum

max: maximum

median: median

mean: mean

iqr: interquartile range

mad: median absolute deviation (see ?MAD)

sd: standard deviation of the mean

se: standard error of the mean

ci: confidence interval of the mean

range: the range = max - min

cv: coefficient of variation, sd/mean

var: variance, sd^2

Examples

# Load data
data("ToothGrowth")

# Descriptive statistics
res <- desc_statby(ToothGrowth, measure.var = "len",
   grps = c("dose", "supp"))
head(res[, 1:10])

#>   dose supp length  min  max median mean iqr mad  sd
#> 1  0.5   OJ     10  8.2 21.5   12.2 13.2 6.5 4.3 4.5
#> 2  0.5   VC     10  4.2 11.5    7.2  8.0 4.9 3.6 2.7
#> 3  1.0   OJ     10 14.5 27.3   23.5 22.7 5.4 3.9 3.9
#> 4  1.0   VC     10 13.6 22.5   16.5 16.8 2.0 1.7 2.5
#> 5  2.0   OJ     10 22.4 30.9   25.9 26.1 2.5 2.1 2.7
#> 6  2.0   VC     10 18.5 33.9   25.9 26.1 5.4 4.6 4.8