Execute ggplot2 functions

A helper function used by ggpubr functions to execute any geom_* functions in ggplot2. Useful only when you want to call a geom_* function without carrying about the arguments to put in aes(). Basic users of ggpubr don't need this function.

geom_exec(geomfunc = NULL, data = NULL, position = NULL, ...)

Arguments

geomfunc

a ggplot2 function (e.g.: geom_point)

data

a data frame to be used for mapping

position

Position adjustment, either as a string, or the result of a call to a position adjustment function.

...

arguments accepeted by the function

Value

return a plot if geomfunc!=Null or a list(option, mapping) if geomfunc = NULL.

Examples

## Not run: ------------------------------------
# ggplot() + geom_exec(geom_point, data = mtcars,
#     x = "mpg", y = "wt", size = "cyl", color = "cyl")
## ---------------------------------------------