Common parameters for ggpubr stat and geom functions

Common parameters for ggpubr stat and geom functions

Arguments

inherit.aes

If FALSE (the default for most ggpubr functions), overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification. Set to TRUE to inherit aesthetics from the parent ggplot layer.